jwt-weak/requirements.txt
Flask~=2.3
PyJWT~=2.9